apps/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pydantic-settings>=2.1.0
python-multipart>=0.0.6
email-validator>=2.0.0
httpx>=0.25.0
boto3>=1.34.0
redis>=5.0.0
pytest>=7.0.0
